- Adjustments for latest bgfx changes

+++ engine/source/plugins/plugins_shared.h
@@ -174,7 +174,7 @@ namespace Plugins
       void (*dbgTextClear)(uint8_t _attr, bool _small); // Defaults: _attr = 0, _small = false
 	   void (*dbgTextPrintf)(uint16_t _x, uint16_t _y, uint8_t _attr, const char* _format, ...);
 
-      void (*setViewClear)(uint8_t _id, uint8_t _flags, uint32_t _rgba, float _depth, uint8_t _stencil); // Defaults: _depth = 1.0, _stencil = 0
+      void (*setViewClear)(uint8_t _id, uint16_t _flags, uint32_t _rgba, float _depth, uint8_t _stencil); // Defaults: _depth = 1.0, _stencil = 0
       void (*setViewRect)(uint8_t _id, uint16_t _x, uint16_t _y, uint16_t _width, uint16_t _height);
       void (*setViewTransform)(uint8_t _id, const void* _view, const void* _projL, uint8_t _flags, const void* _projR); // Defaults: _flags = BGFX_VIEW_STEREO, _projR = NULL
